+++ suite-etl/requirements.txt
@@ -2,5 +2,5 @@ pyiceberg[pandas,pyarrow,polars]>=0.5.0
 boto3>=1.28.0
 aiohttp>=3.8.0
 python-dotenv>=1.0.0 
-fastapi[standard]
-web3
+fastapi[standard]>=0.109.1
+web3>=6.0.0,<7.0.0
